+++ src/webapp-builtins/export/ProcessIdFile.js
@@ -5,7 +5,7 @@ import * as fs from 'node:fs/promises';
 import { pid as processPid } from 'node:process';
 
 import { Threadlet } from '@this/async';
-import { WallClock } from '@this/clocks';
+import { WallClock } from '@this/clocky';
 import { Duration } from '@this/data-values';
 import { Statter } from '@this/fs-util';
 import { ProcessUtil } from '@this/host';